fix(build): portable printf formats for uint64_t on Linux/gcc
The determinism and tsan CI jobs (added with the water/boundary work)
run on ubuntu and had never been exercised — feature-branch pushes do
not trigger code-quality. First run flagged %llu/%llx fed uint64_t,
which is 'long unsigned int' on LP64 Linux, failing -Werror=format.
Cast the 64-bit args to unsigned long long at every %ll site that
lacked it (asset_manager, package_container, asset_types, world,
async_io); the remaining sites already cast.

+++ engine/runtime/asset/src/asset_manager.cpp
@@ -639,8 +639,8 @@ AssetHandle AssetManager::LoadAssetSync(const std::string& assetName) {
     loadedAssetsCount_++;
     totalMemory_ += data->GetPayloadSize();
 
-    NEXT_LOG_INFO("Asset loaded successfully: %s (ID: %llu, size: %zu bytes)",
-                  storageKey.c_str(), id, data->GetPayloadSize());
+    NEXT_LOG_INFO("Asset loaded successfully: %s (ID: %llu, size: %zu bytes)", storageKey.c_str(),
+                  static_cast<unsigned long long>(id), data->GetPayloadSize());
 
     return AssetHandle(id, data.get());
 }
